Final Commit for now
It works.. I took a few liberies on what i added but it should have
everything that you asked for. Also note you need to create new ojbects
with the (+) button in the tableview controller to get the views that
continually drill into themselves. Make sure you tap on those objects.