Work in progress, currently getting the stats generated.
Working on creating "Weather Wonder" app that will use the Core Plot
library to show the charts over the next two weeks for predictions.
This app should be very straight forward as it will have a tab for each
of the informaton feilds. Current status is still getting all the
statistics generated.